FIX JBSEAM-4435 Bad redirect after exception in custom resource
Exception in Seam Resource is Swallowed and Redirection to debug.seam
Does not Work.

In this situation we can not guess the extension used for jsf. (url
pattern or patterns defined in web.xml). So, we redirect to /debug.xhtml
or /error.xhtml directly, and you must configure to process xhtml with
FacesServlet