Clean jar-dependencies folder.
The build phase folder jar-dependencies, used to collect all module
dependencies for use in distribution, never gets cleaned.

At some point someone attempted this as part of the distribution module
but it gets a full lifecycle build in profile distribution and will
clean the folder before it has a chance to use it. The responsibility
then goes to the first module in the normal build process which
populates this folder.